check user auth

+++ src/controllers/contacts.js
@@ -10,8 +10,9 @@ import { parsePaginationParams } from '../utils/parsePaginationParams.js';
 import { parseSortParams } from '../utils/parseSortParams.js';
 import { parseFilterParams } from '../utils/parseFilterParams.js';
 
-export const createContactController = async (req, res) => {
-  const contact = await createContact(req.body);
+export const createContactController = async (req, res, next) => {
+  const reqData = { ...req.body, userId: req.user._id };
+  const contact = await createContact(reqData);
 
   res.status(201).json({
     status: 201,
@@ -31,6 +32,7 @@ export const getContactsController = async (req, res, next) => {
       sortBy,
       sortOrder,
       filter,
+      userId: req.user._id.toString(),
     });
 
     res.json({
@@ -47,6 +49,14 @@ export const getContactByIdController = async (req, res, next) => {
   const { contactId } = req.params;
   const contact = await getContactById(contactId);
 
+  if (contact.userId.toString() !== req.user._id.toString()) {
+    res.status(401).json({
+      status: 401,
+      message: 'You need to have access rights to this contact',
+    });
+    return;
+  }
+
   if (!contact) {
     throw createHttpError(404, 'Contact not found');
   }
@@ -59,6 +69,14 @@ export const getContactByIdController = async (req, res, next) => {
 };
 export const deleteContactController = async (req, res, next) => {
   const { contactId } = req.params;
+  const authContact = await getContactById(contactId);
+  if (authContact.userId.toString() !== req.user._id.toString()) {
+    res.status(401).json({
+      status: 401,
+      message: 'You need to access rights to this contact',
+    });
+    return;
+  }
   const contact = await deleteContact(contactId);
 
   if (!contact) {
@@ -91,7 +109,18 @@ export const upsertContactController = async (req, res, next) => {
 };
 export const patchContactController = async (req, res, next) => {
   const { contactId } = req.params;
-  const result = await updateContact(contactId, req.body);
+  const { body } = req;
+  const authContact = await getContactById(contactId);
+
+  if (authContact.userId.toString() !== req.user._id.toString()) {
+    res.status(401).json({
+      status: 401,
+      message: 'You need to access rights to this contact',
+    });
+    return;
+  }
+  const { result } = await updateContact(contactId, body);
+  console.log(result);
 
   if (!result) {
     next(createHttpError(404, 'Contact not found'));
